refacto: split MailClient class into its separate files.
